requirements.txt
argparse
dlib==19.21.0
h5py==2.10.0
imutils
joblib==0.17.0
jpeg2dct
matplotlib==3.3.2
numpy==1.19.1
opencv-python==4.4.0.44
pandas==0.23.0
Pillow==8.0.0
PyYAML==5.3.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-image
scikit-learn==0.23.2
scikit-video==1.1.11
scipy==1.5.2
seaborn==0.11.0
sklearn==0.0
tensorboard==1.12.2
tensorboardX==2.1
thop
timm
torchstat
torchvision
tqdm==4.51.0
tzlocal